Moved parsing of jx args and define stmts to single location
